requirements.txt
streamlit
python-dotenv
google-generativeai
langchain
langchain-community
faiss-cpu
PyPDF2
tabula-py
pandas
numpy
python-multipart